Create repository and persistance logic

+++ src/main/kotlin/com/example/todo/controllers/TodoController.kt
@@ -12,6 +12,6 @@ class TodoController {
     fun getTodo(@PathVariable(value = "id") id: String): TodoModel {
         val templateTitle = "This is my todo"
         val templateDescription = "This is my todo's description"
-        return TodoModel(id = id, title = templateTitle, description = templateDescription)
+        return TodoModel(id = id.toLong(), title = templateTitle, description = templateDescription)
     }
 }
